+++ api-base/src/app/(rest)/welcome.controller.ts
@@ -1,6 +1,13 @@
-import { Controller, Get, Post, ZanixController } from '@zanix/server'
+import {
+  Controller,
+  Get,
+  Interceptor,
+  Post,
+  ZanixController
+} from '@zanix/server'
 import { ExampleService } from 'services/example.service'
 import { ExampleRTO } from './rtos'
+import { exampleInterceptor } from 'middlewares'
 
 /**
  * @Controller
@@ -14,17 +21,46 @@ import { ExampleRTO } from './rtos'
 
 @Controller('welcome', ExampleService)
 export class WelcomeController extends ZanixController<ExampleService> {
+  /**
+   * @name welcome
+   * @summary GET api/welcome route
+   * @returns {HttpResponse}
+   */
   @Get()
   public async welcome() {
     return {
-      message: 'Welcome ZanixJS Server'
+      message: 'Welcome to ZanixJS Server'
     }
   }
 
+  /**
+   * @name welcomeIntercepted
+   * @summary GET api/welcome/intercepted route
+   * @description Shows an option to intercept response using @Interceptor decorator
+   * @returns {HttpResponse}
+   */
+  @Get('intercepted')
+  @Interceptor(exampleInterceptor)
+  public async welcomeIntercepted() {
+    return {
+      message: 'Welcome to ZanixJS Server'
+    }
+  }
+
+  /**
+   * @name post
+   * @summary POST api/welcome route
+   * @description Request object validation {ExampleRTO} and processing
+   * @param payload
+   * @returns
+   */
   @Post(ExampleRTO)
   public async post(payload: Welcome) {
     const data = await this.service.welcome(payload)
 
-    return { message: 'Request processed', data }
+    return {
+      message: 'Request processed',
+      data: { data, piped: this.context.req.body.piped }
+    }
   }
 }